refactor(home): 홈 섹션 도메인 시맨틱 정리 (CCTV/Website 제거, 4섹션 키 리네임·카피 갱신)
건설 현장 본연의 도메인에 맞춰 홈 섹션을 재정리. CCTV 위험감시·웹사이트 구축 섹션은 제거하고, dashboard 외 approval/employee/aiReport 키를 의미가 일치하는 blueprint/documents/workReports로 리네임. 카피와 이미지도 자금 현황·공정표, 도면, 문서, 작업일보 흐름에 맞춰 ko/en/ja 동시 갱신.
