+++ messages/ja.json
@@ -146,46 +146,31 @@
     },
     "dashboard": {
       "subtitle": "プロジェクト状況を一目で可視化",
-      "title": "全社ダッシュボード",
-      "body1": "工程、日程、資金状況、課題など、プロジェクトの主要指標と流れを一目で把握できます。",
-      "body2": "迅速な理解を通じて、円滑な意思決定を支援します。",
-      "imageAlt": "全社ダッシュボード画面キャプチャ"
+      "title": "資金状況・工程表",
+      "body1": "プロジェクト文書データをもとに資金フローを可視化し、意思決定を支援します。",
+      "body2": "工種・工程作業を中心とした体系的な工程表で、主要なマイルストーンから残作業まで一目で把握できます。",
+      "imageAlt": "資金状況・工程表画面キャプチャ"
     },
-    "approval": {
-      "subtitle": "文書をオンラインで作成・確認・承認",
-      "title": "電子決裁システム",
-      "body1": "文書をオンラインで処理し、決裁手続きを迅速に進められます。",
-      "body2": "決裁状況を一か所で確認し、業務を体系的に管理できます。",
-      "imageAlt": "電子決裁システム画面キャプチャ"
+    "blueprint": {
+      "subtitle": "位置情報をもとに現場情報を照会する",
+      "title": "プロジェクト図面管理",
+      "body1": "図面のバージョンや変更履歴を追跡し、位置タグに基づく施工記録・瑕疵管理・メッセージを送信します。",
+      "body2": "領域別のマーカー移動と簡単な照会を統合的に支援するプロジェクト図面管理機能です。",
+      "imageAlt": "プロジェクト図面管理画面キャプチャ"
     },
-    "employee": {
-      "subtitle": "社員情報を統合して管理",
-      "title": "社員管理システム",
-      "body1": "社員情報を一か所に集約し、勤怠や人事状況を簡単に確認できます。",
-      "body2": "必要なデータを体系的に整理し、効率的な人材管理を支援します。",
-      "imageAlt": "社員管理システム画面"
+    "documents": {
+      "subtitle": "煩雑な手入力なしで埋める",
+      "title": "プロジェクト文書管理",
+      "body1": "工事台帳や出来高請求書など重要な文書を体系的に管理します。",
+      "body2": "AI技術で書類ベースのデータ入力を自動化し、文書業務の効率性と正確性を最大化します。",
+      "imageAlt": "プロジェクト文書管理画面"
     },
-    "aiReport": {
+    "workReports": {
       "subtitle": "蓄積データから自動作成",
-      "title": "AIレポート",
+      "title": "作業日報自動生成",
       "body1": "蓄積データを分析し、AIがレポートを自動で完成させます。",
       "body2": "項目設定やファイル添付により、内容と様式をカスタマイズできます。",
-      "imageAlt": "AIレポート自動生成画面キャプチャ"
-    },
-    "cctv": {
-      "subtitle": "AIが自動で分析",
-      "title": "CCTVリアルタイム危険検知",
-      "body1": "AIが複数のCCTV映像をリアルタイムで分析し、ヘルメット未着用や危険区域侵入などの異常を迅速に検出します。",
-      "body2": "過去映像とイベント履歴を確認し、現場の安全状態を把握できます。",
-      "imageAlt": "CCTVリアルタイム危険検知画面"
-    },
-    "website": {
-      "badge": "PREMIUM PLAN +",
-      "subtitle": "企業を紹介し、魅力を伝える",
-      "title": "Webサイト構築",
-      "body1": "Proプラン以上のお客様向けに、企業ブランディングと要件を反映したWebサイトを制作します。",
-      "body2": "専門的なデザインとWeb構築で企業を効果的にPRします。",
-      "imageAlt": "Webサイト構築サービス紹介画面"
+      "imageAlt": "作業日報自動生成画面キャプチャ"
     },
     "contactSection": {
       "headingLine1": "RENAME DP導入のご相談、",
