+++ messages/ko.json
@@ -146,47 +146,33 @@
     },
     "dashboard": {
       "subtitle": "프로젝트 상태를 한눈에 보여주는",
-      "title": "전사 현황판",
-      "body1": "공정, 일정, 자금 현황, 이슈 등 프로젝트 핵심 지표와 흐름을 한눈에 파악할 수 있습니다.",
-      "body2": "빠른 이해를 통해 원활한 의사 결정을 돕습니다.",
-      "imageAlt": "전사 현황판 대시보드 화면 캡처 이미지"
+      "title": "자금 현황·공정표",
+      "body1": "프로젝트 문서 데이터를 기반으로 자금 흐름을 시각화하여 의사 결정을 돕습니다.",
+      "body2": "공종·공정 작업 중심의 체계적인 공정표를 통해 주요 공지부터 잔여 업무까지 한눈에 파악합니다.",
+      "imageAlt": "자금 현황·공정표 화면 캡처 이미지"
     },
-    "approval": {
-      "subtitle": "문서를 온라인으로 작성·검토·승인하는",
-      "title": "전자 결재 시스템",
-      "body1": "문서를 온라인으로 처리해 결재 절차를 빠르게 진행할 수 있습니다.",
-      "body2": "결재 현황을 한곳에서 확인하며 업무를 체계적으로 관리할 수 있습니다.",
-      "imageAlt": "전자 결재 시스템 문서 작성 화면 캡처 이미지"
+    "blueprint": {
+      "subtitle": "위치 기반으로 현장 정보를 조회하는",
+      "title": "프로젝트 도면 관리",
+      "body1": "도면 버전 및 변경 사항 추적, 위치 태그 기반 시공 기록·하자 관리·메시지를 전송합니다.",
+      "body2": "영역별 마커 이동 및 간편 조회를 통합 지원하는 프로젝트 도면 관리 기능입니다.",
+      "imageAlt": "프로젝트 도면 관리 화면 캡처 이미지"
     },
-    "employee": {
-      "subtitle": "직원 정보를 통합해 관리하는",
-      "title": "직원 관리 시스템",
-      "body1": "직원 정보를 한곳에 모아 근태와 인사 현황을 손쉽게 확인할 수 있습니다.",
-      "body2": "필요한 데이터를 체계적으로 정리해 효율적인 인력 관리를 지원합니다.",
-      "imageAlt": "직원 관리 시스템 화면"
+    "documents": {
+      "subtitle": "번거로운 수기 입력 없이 채우는",
+      "title": "프로젝트 문서 관리",
+      "body1": "공사대장, 기성 청구서 등 중요 문서를 체계적으로 관리합니다.",
+      "body2": "AI 기술로 서류 기반 데이터 입력을 자동화하여 문서 작업의 효율성과 정확도를 극대화합니다.",
+      "imageAlt": "프로젝트 문서 관리 화면"
     },
-    "aiReport": {
+    "workReports": {
       "subtitle": "적재된 데이터로 자동으로 작성되는",
-      "title": "AI 보고서",
+      "title": "작업일보 자동 생성",
       "body1": "적재된 데이터를 분석해 AI가 자동으로 보고서를 완성합니다.",
       "body2": "사용자가 항목을 직접 설정하거나 파일을 첨부해 내용과 양식을 맞춤화할 수 있습니다.",
-      "imageAlt": "AI 보고서 자동 생성 화면 캡처 이미지"
+      "imageAlt": "작업일보 자동 생성 화면 캡처 이미지"
     },
-    "cctv": {
-      "subtitle": "AI가 자동으로 분석하는",
-      "title": "CCTV 실시간 위험 감지",
-      "body1": "AI가 여러 CCTV 영상을 실시간으로 분석해 안전모 미착용, 위험 구역 침입 등 다양한 이상 상황을 빠르게 찾아냅니다.",
-      "body2": "과거 영상과 이벤트 기록을 조회해 현장 안전 상태를 쉽게 파악할 수 있습니다.",
-      "imageAlt": "CCTV 실시간 위험 감지 화면"
-    },
-    "website": {
-      "badge": "PREMIUM PLAN +",
-      "subtitle": "기업을 소개하고 홍보하는",
-      "title": "웹사이트 구축",
-      "body1": "Pro 플랜 이상 고객을 대상으로 기업 브랜딩과 요청 내용을 반영한 웹사이트를 제작합니다.",
-      "body2": "전문 디자인과 웹 구축으로 기업을 효과적으로 홍보합니다.",
-      "imageAlt": "웹사이트 구축 서비스를 소개하는 캡처 화면"
-    },
+    
     "contactSection": {
       "headingLine1": "RENAME DP 도입 상담,",
       "headingLine2": "바로 문의하세요",
