--- src/features/home/components/HomeCctvSection/HomeCctvSection.module.css
@@ -1,157 +1,0 @@
-/* src/features/home/components/HomeCctvSection/HomeCctvSection.module.css */
-
-.section {
-  background: #eef3fb;
-}
-
-.container {
-  display: flex;
-  flex-direction: row;
-  gap: 24px;
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 80px 64px;
-  align-items: center;
-}
-
-/* 텍스트 컬럼 */
-.textColumn {
-  flex: 1 0 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-width: 0;
-}
-
-.textBlock {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  width: 446px;
-  max-width: 100%;
-}
-
-.headerContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 0;
-}
-
-.subtitle {
-  font-family: var(--font-noto-sans-kr), 'Noto Sans KR', sans-serif;
-  font-size: 28px;
-  font-weight: 600;
-  line-height: 48px;
-  letter-spacing: 0px;
-  color: var(--color-heading);
-  word-break: keep-all;
-  margin: 0;
-}
-
-.title {
-  font-family: var(--font-noto-sans-kr), 'Noto Sans KR', sans-serif;
-  font-size: 40px;
-  font-weight: 600;
-  line-height: 32px;
-  letter-spacing: 0px;
-  color: var(--color-accent);
-  word-break: keep-all;
-  margin: 0;
-}
-
-.body {
-  font-family: var(--font-noto-sans-kr), 'Noto Sans KR', sans-serif;
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 1.45;
-  letter-spacing: -0.12px;
-  color: var(--color-body);
-  word-break: keep-all;
-  margin: 0;
-}
-
-/* 이미지 컬럼 */
-.imageColumn {
-  flex: 1 0 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-width: 0;
-}
-
-.imageWrapper {
-  position: relative;
-  width: 100%;
-  aspect-ratio: 1429 / 1060;
-}
-
-.image {
-  object-fit: cover;
-}
-
-/* 태블릿 */
-@media (min-width: 800px) and (max-width: 1279px) {
-  .container {
-    padding: 80px 48px;
-  }
-
-  .textBlock {
-    width: 345px;
-  }
-
-  .subtitle {
-    font-size: 18px;
-    line-height: 32px;
-  }
-
-  .title {
-    font-size: 28px;
-    font-weight: 700;
-    line-height: 1.5;
-    letter-spacing: -2px;
-  }
-
-  .body {
-    font-size: 16px;
-  }
-}
-
-/* 모바일 */
-@media (max-width: 799px) {
-  .container {
-    flex-direction: column;
-    gap: 32px;
-    padding: 48px 20px;
-  }
-
-  .textColumn {
-    width: 100%;
-    align-items: flex-start;
-    justify-content: flex-start;
-  }
-
-  .textBlock {
-    width: 280px;
-    max-width: 100%;
-  }
-
-  .subtitle {
-    font-size: 18px;
-    line-height: 32px;
-  }
-
-  .title {
-    font-size: 28px;
-    font-weight: 700;
-    line-height: 1.5;
-    letter-spacing: -2px;
-  }
-
-  .body {
-    font-size: 16px;
-  }
-
-  .imageColumn {
-    width: 100%;
-  }
-}
