+++ src/features/home/components/HomeDocumentsSection/HomeDocumentsSection.module.css
@@ -1,4 +1,4 @@
-/* src/features/home/components/HomeApprovalSection/HomeApprovalSection.module.css */
+/* src/features/home/components/HomeDocumentsSection/HomeDocumentsSection.module.css */
 
 .section {
   background: #eef3fb;
@@ -14,85 +14,89 @@
   align-items: center;
 }
 
-/* 이미지가 LEFT, 텍스트가 RIGHT */
-.imageColumn {
-  flex: 1;
+/* 텍스트 컬럼 */
+.textColumn {
+  flex: 1 0 0;
   display: flex;
   align-items: center;
   justify-content: center;
+  min-width: 0;
+}
+
+.textBlock {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  width: 437px;
+  max-width: 100%;
+}
+
+.headerContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 0;
+}
+
+.subtitle {
+  font-family: var(--font-noto-sans-kr), 'Noto Sans KR', sans-serif;
+  font-size: 28px;
+  font-weight: 600;
+  line-height: 48px;
+  letter-spacing: 0px;
+  color: #1e2939;
+  word-break: keep-all;
+  margin: 0;
+}
+
+.title {
+  font-family: var(--font-noto-sans-kr), 'Noto Sans KR', sans-serif;
+  font-size: 40px;
+  font-weight: 600;
+  line-height: 32px;
+  letter-spacing: 0px;
+  color: #2b7fff;
+  word-break: keep-all;
+  margin: 0;
+}
+
+.body {
+  font-family: var(--font-noto-sans-kr), 'Noto Sans KR', sans-serif;
+  font-size: 16px;
+  font-weight: 400;
+  line-height: 1.45;
+  letter-spacing: -0.12px;
+  color: rgba(0, 0, 0, 0.55);
+  word-break: keep-all;
+  margin: 0;
+}
+
+/* 이미지 컬럼 */
+.imageColumn {
+  flex: 1 0 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  min-width: 0;
 }
 
 .imageWrapper {
   position: relative;
-  aspect-ratio: 1996 / 1388;
   width: 100%;
+  aspect-ratio: 2433 / 1239;
 }
 
 .image {
   object-fit: cover;
 }
 
-.textColumn {
-  flex: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  align-self: stretch;
-}
-
-.textBlock {
-  width: 451px;
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  align-items: flex-start;
-}
-
-.headerContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
-
-.subtitle {
-  font-size: 28px;
-  font-weight: 600;
-  color: var(--color-heading);
-  line-height: 48px;
-  font-family: var(--font-noto-sans-kr), sans-serif;
-  word-break: keep-all;
-  margin: 0;
-}
-
-.title {
-  font-size: 40px;
-  font-weight: 600;
-  color: var(--color-accent);
-  line-height: 32px;
-  font-family: var(--font-noto-sans-kr), sans-serif;
-  word-break: keep-all;
-  margin: 0;
-}
-
-.body {
-  font-size: 24px;
-  font-weight: 400;
-  color: var(--color-body);
-  line-height: 1.45;
-  letter-spacing: -0.12px;
-  font-family: var(--font-noto-sans-kr), sans-serif;
-  word-break: keep-all;
-  margin: 0;
-}
-
-/* Tablet: 800px ~ 1279px */
+/* 태블릿 */
 @media (min-width: 800px) and (max-width: 1279px) {
   .container {
     padding: 80px 48px;
   }
 
   .textBlock {
-    width: 100%;
+    width: 320px;
   }
 
   .subtitle {
@@ -102,17 +106,12 @@
 
   .title {
     font-size: 28px;
-    font-weight: 700;
-    letter-spacing: -2px;
     line-height: 1.5;
-  }
-
-  .body {
-    font-size: 16px;
+    letter-spacing: -2px;
   }
 }
 
-/* Mobile: ≤799px */
+/* 모바일 */
 @media (max-width: 799px) {
   .container {
     flex-direction: column;
@@ -120,20 +119,13 @@
     padding: 48px 20px;
   }
 
-  .imageColumn {
-    width: 100%;
-    order: 2;
-  }
-
   .textColumn {
     width: 100%;
-    align-items: flex-start;
-    justify-content: flex-start;
-    order: 1;
   }
 
   .textBlock {
-    width: 100%;
+    width: 280px;
+    max-width: 100%;
   }
 
   .subtitle {
@@ -143,12 +135,11 @@
 
   .title {
     font-size: 28px;
-    font-weight: 700;
-    letter-spacing: -2px;
     line-height: 1.5;
+    letter-spacing: -2px;
   }
 
-  .body {
-    font-size: 16px;
+  .imageColumn {
+    width: 100%;
   }
 }
