+++ src/features/home/components/HomeWorkReportsSection/HomeWorkReportsSection.module.css
@@ -1,4 +1,4 @@
-/* src/features/home/components/HomeAiReportSection/HomeAiReportSection.module.css */
+/* src/features/home/components/HomeWorkReportsSection/HomeWorkReportsSection.module.css */
 
 .section {
   background: #eef3fb;
